refactor: 피드 api 통합
/api/feed에 GET요청시 메인 페이지 피드 정보를 모두 불러오고, POST 요청시 피드를 업로드 할 수 있다.
