+++ firebase-practice/pages/api/feed.ts
@@ -1,13 +1,18 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import { DBService } from "@FireBase"
 import { FeedData } from "backend/dto"
-import { doc, getDoc } from "firebase/firestore"
+import { arrayUnion, doc, getDoc, setDoc, updateDoc } from "firebase/firestore"
 import type { NextApiRequest, NextApiResponse } from "next"
 
 /**
- * request url : /api/feed
  * method : GET
+ * request url : /api/feed
  * response : 메인 페이지의 피드 정보들
+ *
+ * method : POST
+ * request url : /api/feed
+ * request body : FeedData
+ * response: 성공시 "Success", 실패시 "Fail"
  */
 
 export default async function getFeed(
@@ -26,5 +31,54 @@ export default async function getFeed(
     } else {
       res.status(500).json("Fail")
     }
+  } else if (req.method === "POST") {
+    const {
+      imageUrl,
+      desc,
+      location,
+      isPrivate,
+      storageId,
+      creator,
+      uploadTime,
+    } = req.body as FeedData
+
+    const firestoreAllRef = doc(DBService, "mainPage", `userFeedDataAll`)
+    const firestorePersonalRef = doc(DBService, "users", `${creator}`)
+
+    const feed: FeedData = {
+      imageUrl: imageUrl,
+      desc: desc,
+      location: location,
+      isPrivate: isPrivate,
+      storageId: storageId,
+      creator: creator,
+      uploadTime: uploadTime,
+    }
+
+    await updateDoc(firestoreAllRef, {
+      feed: arrayUnion(feed),
+    }).catch(async (error) => {
+      if (error.code === "not-found") {
+        await setDoc(firestoreAllRef, {
+          feed: [feed],
+        })
+      } else {
+        res.status(500).json(error.code)
+      }
+    })
+
+    await updateDoc(firestorePersonalRef, {
+      feed: arrayUnion(feed),
+    }).catch(async (error) => {
+      if (error.code === "not-found") {
+        await setDoc(firestorePersonalRef, {
+          feed: [feed],
+        })
+      } else {
+        res.status(500).json(error.code)
+      }
+    })
+
+    res.status(200).json("Success")
   }
 }
